tetragon: Extract execve event initialization
Adding execve_event_init function that initialize exeve event,
no functional change.
