+++ bpf/process/bpf_execve_event.h
@@ -4,6 +4,7 @@
 #ifndef __BPF_EXECVE_EVENT_H__
 #define __BPF_EXECVE_EVENT_H__
 
+#include "bpf_mbset.h"
 #include "data_event.h"
 
 struct {
@@ -211,6 +212,73 @@ read_execve_shared_info(void *ctx, struct msg_process *p, __u64 pid)
 	p->i_ino = info->i_ino;
 	p->i_nlink = info->i_nlink;
 	execve_joined_info_map_clear(pid);
+}
+
+FUNC_LOCAL void
+execve_event_init(struct bpf_raw_tracepoint_args *ctx,
+		  struct msg_execve_event *event)
+{
+	struct task_struct *task = (struct task_struct *)get_current_task();
+	struct linux_binprm *bprm = (struct linux_binprm *)ctx->args[2];
+	struct execve_map_value *parent;
+	struct msg_process *p;
+	char *filename;
+	__u64 pid;
+
+	pid = get_current_pid_tgid();
+	parent = event_find_parent();
+	if (parent) {
+		event->parent = parent->key;
+		update_mb_task(parent, &parent->bin);
+		event->parent_flags = 0;
+	} else {
+		event_minimal_parent(event, task);
+	}
+
+	p = &event->process;
+	p->flags = EVENT_EXECVE;
+	p->size_path = 0;
+	p->size_args = 0;
+	p->size_cwd = 0;
+	p->size_envs = 0;
+
+	/**
+	 * Per thread tracking rules TID == PID :
+	 *  At exec all threads other than the calling one are destroyed, so
+	 *  current becomes the new thread leader since we hook late during
+	 *  execve.
+	 */
+	p->pid = pid >> 32;
+	p->tid = (__u32)pid;
+	p->nspid = get_task_pid_vnr_curr();
+	p->ktime = tg_get_ktime();
+	p->size = offsetof(struct msg_process, args);
+	p->auid = get_auid();
+	read_execve_shared_info(ctx, p, pid);
+
+	probe_read(&filename, sizeof(filename), _(&bprm->filename));
+	p->size += read_path(ctx, event, filename);
+	p->size += read_args(ctx, event);
+	p->size += read_cwd(ctx, p);
+	p->size += read_envs(ctx, event);
+
+	event->common.op = MSG_OP_EXECVE;
+	event->common.flags = 0;
+	event->common.ktime = p->ktime;
+	event->common.size = offsetof(struct msg_execve_event, process) + p->size;
+
+	get_current_subj_creds(&event->creds, task);
+	/**
+	 * Instead of showing the task owner, we want to display the effective
+	 * uid that is used to calculate the privileges of current task when
+	 * acting upon other objects. This allows to be compatible with the 'ps'
+	 * tool that reports snapshot of current processes.
+	 */
+	p->uid = event->creds.euid;
+	get_namespaces(&event->ns, task);
+
+	// Zero the cleanup key to prevent user space confusion.
+	event->cleanup_key = (struct msg_execve_key){ 0 };
 }
 
 #endif /* __BPF_EXECVE_EVENT_H__ */
